+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Allow accessing `state.value` after destroying.
+- Added toolbar config `collapsed` to set the default collapsed state of toolbar.
 
 ## 0.3.14